Rich text: move key handlers to ref callbacks
